+++ src/components/reply-modal/reply-modal.tsx
@@ -63,6 +63,16 @@ const ReplyModal = ({ closeModal, parentCid, scrollY }: ReplyModalProps) => {
     }
   }, [isMobile, scrollY]);
 
+  const parentCidRef = useRef<HTMLSpanElement>(null);
+  const parentCidWrapperRef = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    if (parentCidRef.current && parentCidRef.current) {
+      const cidWidth = parentCidRef.current.offsetWidth;
+      parentCidRef.current.style.width = `${cidWidth}px`;
+    }
+  }, [parentCid]);
+
   const modalContent = (
     <div className={styles.container} ref={nodeRef}>
       <div className={`replyModalHandle ${styles.title}`}>
@@ -88,10 +98,12 @@ const ReplyModal = ({ closeModal, parentCid, scrollY }: ReplyModalProps) => {
         <div className={styles.link}>
           <input type='text' ref={urlRef} placeholder={_.capitalize(t('link'))} onChange={(e) => setContent.link(e.target.value)} />
         </div>
-        <div className={styles.parentCid}>
-          <input type='text' readOnly value={`c/${parentCid && Plebbit.getShortCid(parentCid)}`} />
-        </div>
         <div className={styles.content}>
+          <span className={styles.parentCidWrapper} ref={parentCidWrapperRef}>
+            <span className={styles.parentCid} ref={parentCidRef}>
+              {`c/${parentCid && Plebbit.getShortCid(parentCid)}`}
+            </span>
+          </span>
           <textarea
             cols={48}
             rows={4}